requirements.txt
Flask>=2.0
transformers>=4.0
python-dotenv>=0.19.0